fixed deleting from cart

+++ course-work-client/src/app/modules/cart/book-cart/book-cart.component.ts
@@ -1,7 +1,8 @@
-import {Component, Input, OnInit} from '@angular/core';
+import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
 import {Book} from '../../../models/common/book.model';
 import {TokenStorageService} from '../../../core/services/token-storage/token-storage.service';
 import {CartService} from '../../../core/services/cart/cart.service';
+import {count} from 'rxjs/operators';
 
 @Component({
   selector: 'app-book-cart',
@@ -12,6 +13,8 @@ export class BookCartComponent implements OnInit {
   @Input()
   bookEntity: Book;
 
+  @Output()
+  isReloadEvent = new EventEmitter<boolean>();
   constructor(
     private token: TokenStorageService,
     private cartService: CartService
@@ -22,7 +25,15 @@ export class BookCartComponent implements OnInit {
   }
 
   onDelete(): void {
-    this.cartService.deleteBookFromCart(this.bookEntity.id, this.token.getUsername());
+    this.cartService.getCountOBook(this.bookEntity.id, this.token.getUsername()).subscribe((countOfBook: number) => {
+      if (countOfBook === 1) {
+        console.log('here');
+        this.cartService.deleteBookFromCart(this.bookEntity.id, this.token.getUsername());
+        this.isReloadEvent.emit(true);
+        } else {
+        this.cartService.deleteBookFromCart(this.bookEntity.id, this.token.getUsername());
+      }
+    });
   }
 
 }
